[Feat] Error 추가 & 로직 수정
User-Soundtrack 불일치 에러
음원 미존재 에러
추가에 따른 , delete 와 update 로직 수정

+++ src/services/soundtrack.service.ts
@@ -3,8 +3,10 @@ import { Inject, Injectable } from "@nestjs/common";
 //import { SoundtrackRepository } from "@repositories/soundtrack.repository";
 import { CreateSoundtrackRequestDto } from "src/dtos/soundtrack/create-soundtrack-request.dto";
 import { UpdateSoundtrackRequestDto } from "src/dtos/soundtrack/update-soundtrack-request.dto";
+import { NotExistedSoundtrack } from "src/errors/soundtrack/not-existed-soundtrack.error";
+import { NotMatchedUser } from "src/errors/soundtrack/not-matched-user.error";
 import { Repository } from "typeorm";
-
+import typia from "typia";
 
 
 
@@ -18,18 +20,31 @@ export class SoundtrackService{
 
             return await this.soundtrackRepository.save({creatorId:email,...createSoundtrackRequestDto})
         }catch(err){
-            console.log(err);
             return true
         }
     }
-    async update(trackId : string,updateSoundtrackRequestDto:UpdateSoundtrackRequestDto){
+    async update(email:string,trackId : number,updateSoundtrackRequestDto:UpdateSoundtrackRequestDto){
+        const record = await this.soundtrackRepository.findOne({where:{id:trackId}});
+        if(!record)
+            throw typia.random<NotExistedSoundtrack>();
+        if (record.creatorId!==email)
+            throw typia.random<NotMatchedUser>();
         return await this.soundtrackRepository.update(trackId,updateSoundtrackRequestDto);
-    }
-    async delete(trackId : string){
+    }   
+    async delete(email:string,trackId : number){
+        const record = await this.soundtrackRepository.findOne({where:{id:trackId}});
+        if(!record)
+            throw typia.random<NotExistedSoundtrack>();
+        if (record.creatorId!==email)
+            throw typia.random<NotMatchedUser>();
+
         return await this.soundtrackRepository.delete(trackId);
     }
-    async getOne(trackId : string){
-        return await this.soundtrackRepository.findOne({where:{id:trackId}});
-    }
+    async getOne(trackId : number){
+        const record = await this.soundtrackRepository.findOne({where:{id:trackId}});
+        if(!record)
+            throw typia.random<NotExistedSoundtrack>();
+        return record;
+        }
     async getMany(){}
 }
